pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "fdray"
version = "0.0.1"
description = "Finite difference ray tracing"
readme = "README.md"
license = { file = "LICENSE" }
authors = [{ name = "daizutabi", email = "daizutabi@gmail.com" }]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
]
requires-python = ">=3.10"
dependencies = ["numpy>=2", "moviepy>=2", "pillow>=10"]

[project.urls]
Documentation = "https://github.com/daizutabi/fdray"
Source = "https://github.com/daizutabi/fdray"
Issues = "https://github.com/daizutabi/fdray/issues"

[dependency-groups]
dev = [
  "pytest-clarity>=1",
  "pytest-cov>=6",
  "pytest-randomly>=3.16",
  "pytest-xdist>=3.6",
]

[tool.pytest.ini_options]
addopts = ["--cov=fdray", "--cov-report=lcov:lcov.info", "--doctest-modules"]

[tool.coverage.report]
exclude_lines = ["no cov", "raise NotImplementedError", "if TYPE_CHECKING:"]
skip_covered = true

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["ALL"]
unfixable = ["F401"]
ignore = [
  "A005",
  "ANN002",
  "ANN003",
  "ANN401",
  "ARG002",
  "ARG003",
  "B904",
  "D",
  "D105",
  "D203",
  "D213",
  "EM101",
  "FBT001",
  "PGH003",
  "S603",
  "TRY003",
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = [
  "A001",
  "ANN",
  "ARG",
  "D",
  "FBT",
  "PD",
  "PLR",
  "PT",
  "S",
  "SIM117",
  "SLF",
]

[tool.pyright]
include = ["src", "tests"]
strictDictionaryInference = true
strictListInference = true
strictSetInference = true
